--- a/.github/CHANGELOG.md
+++ b/.github/CHANGELOG.md
@@ -17,13 +17,58 @@ Developers should add to the **Unreleased** section once they feel their branch 
 - Added animated emoji for non Nitro user
 - Add weather, cowsay and insult feature
 
-## \[Unreleased\]
+### Disabled
 
-### Added
+These 3 changes are based on Discord webhooks, and has been temporarily disabled for this
+release because Discord API has changed by the time this was released.
 
 - Added embed message function when user paste in a URL of another message
 - Added poll function when user want to create a simple poll
 - The embed message and poll functions are currently disabled because Discord has updated their APIs.
+
+## 1.0.1
+
+### Added
+
+- Add GHCR path to the package name for deployment
+- Add a CI/CD Pipeline to deploy the service
+- Enable Pipeline to build docker image
+
+### Changed
+
+- Add MSW package to simulate an endpoint in running node-fetch-related tests
+  and remove node-fetch mocks in said test
+
+## 1.1.0
+
+### Added
+
+- Update DiscordJS to v13 to fit with the new Discord API
+- Add Guild Intent to Discord client config
+- Revert the disabling of webhook related commands
+
+## 1.1.1
+
+### Changed
+
+- Replace all `mocked` imports from ts-jest with `jest.mocked`
+
+## 1.1.2
+
+### Changed
+
+- Replace all calls to the faker package with `@faker-js/faker`
+
+## 1.2.0
+
+### Changed
+
+- Add a webpack and esbuild build process into our build and tests pipeline.
+  - The tests will now be run with esbuild jest, making it run a lot faster
+  - The package will now be bundled using webpack and esbuild, making it fast
+  and also reducing the size of our deployment package.
+
+## \[Unreleased\]
 
 [//]: # (Template:)