--- a/docker-compose.override.yml
+++ b/docker-compose.override.yml
@@ -8,7 +8,7 @@ services:
     depends_on:
       - db
     env_file: ".env"
-    image: ghcr.io/viet-aus-it/vait-discord-bot:development
+    image: ghcr.io/viet-aus-it/discord-bot:development
     restart: unless-stopped
     volumes:
       - .:/src