--- a/package.json
+++ b/package.json
@@ -20,8 +20,7 @@
     "prepare": "husky install",
     "clean:all": "npm run prebuild && npm run pretest",
     "prebuild": "rimraf build",
-    "build": "tsc --project tsconfig.build.json",
-    "build:watch": "tsc -w --project tsconfig.build.json",
+    "build": "webpack",
     "lint": "esw src --ext .ts",
     "lint:fix": "npm run lint -- --fix",
     "lint:watch": "npm run lint -- --watch",
@@ -31,7 +30,6 @@
     "test": "jest",
     "test:silent": "jest --silent",
     "test:watch": "npm run test -- --watchAll",
-    "start:script": "ts-node-dev --poll --transpile-only --no-notify ./src/index.ts",
     "start": "concurrently npm:start:script npm:lint:watch",
     "serve": "node build/index.js",
     "prisma:migrate": "prisma migrate dev",
@@ -53,6 +51,8 @@
     "@types/node": "^16.11.12",
     "@types/node-fetch": "^2.5.10",
     "@types/ramda": "^0.27.40",
+    "@types/webpack": "^5.28.0",
+    "@types/webpack-node-externals": "^2.5.3",
     "@typescript-eslint/eslint-plugin": "^5.7.0",
     "@typescript-eslint/parser": "^5.7.0",
     "concurrently": "^6.4.0",
@@ -73,8 +73,11 @@
     "prisma": "^3.6.0",
     "rimraf": "^3.0.2",
     "ts-jest": "^27.1.1",
+    "ts-loader": "^9.2.6",
     "ts-node": "^10.4.0",
-    "ts-node-dev": "^1.1.8",
-    "typescript": "^4.5.4"
+    "typescript": "^4.5.4",
+    "webpack": "^5.66.0",
+    "webpack-cli": "^4.9.1",
+    "webpack-node-externals": "^3.0.0"
   }
 }