--- a/package.json
+++ b/package.json
@@ -32,6 +32,7 @@
     "pretest": "rimraf coverage",
     "test": "jest",
     "test:silent": "jest --silent",
+    "test:typecheck": "concurrently npm:test:silent npm:typecheck",
     "test:watch": "npm run test -- --watchAll",
     "serve": "nodemon",
     "start": "concurrently npm:build:watch npm:serve",
@@ -47,6 +48,7 @@
     "zod": "^3.11.6"
   },
   "devDependencies": {
+    "@babel/preset-typescript": "^7.16.7",
     "@faker-js/faker": "^5.5.3",
     "@jest/types": "^27.4.2",
     "@types/faker": "^5.5.9",
@@ -61,6 +63,7 @@
     "concurrently": "^6.4.0",
     "copy-webpack-plugin": "^10.2.0",
     "esbuild": "^0.14.11",
+    "esbuild-jest": "^0.5.0",
     "esbuild-loader": "^2.18.0",
     "eslint": "^8.4.1",
     "eslint-config-airbnb-base": "^15.0.0",