--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@viet-aus-it/discord-bot",
-  "version": "1.0.0",
+  "version": "1.2.0",
   "description": "Discord chat bot for VAIT",
   "main": "build/index.js",
   "engines": {