--- a/webpack.config.ts
+++ b/webpack.config.ts
@@ -1,0 +1,38 @@
+import path from 'path';
+import { Configuration } from 'webpack';
+import nodeExternals from 'webpack-node-externals';
+
+const isProductionBuild = () => process.env.NODE_ENV === 'production';
+
+const tsconfigFile = isProductionBuild()
+  ? 'tsconfig.json'
+  : 'tsconfig.build.json';
+
+const config: Configuration = {
+  mode: isProductionBuild() ? 'production' : 'development',
+  target: 'node',
+  devtool: isProductionBuild() ? 'eval-source-map' : 'source-map',
+  entry: path.resolve(__dirname, 'src', 'index.ts'),
+  resolve: {
+    extensions: ['.ts', '.js'],
+  },
+  module: {
+    rules: [
+      {
+        test: /\.ts$/,
+        loader: 'ts-loader',
+        options: {
+          configFile: tsconfigFile,
+        },
+      },
+    ],
+  },
+  externalsPresets: { node: true },
+  externals: [nodeExternals()],
+  output: {
+    path: path.resolve(__dirname, 'build'),
+    filename: 'index.js',
+  },
+};
+
+export default config;