Integrate a bundler into this project, which will help with the deployed image size, as well as making the test runs faster.

## Description
<!--- Describe your changes in detail -->
This PR will integrate Webpack and ESBuild into this project, and trying to achieve 2 things:

- With Webpack, it will try to bundle, minify and tree shake all of the dead code in the app and dependencies, making the deployed image smaller. (reducing from 426MB to 237MB)
  - This is with plain TypeScript
  ![Screen Shot 2022-01-15 at 11 52 38 pm](https://user-images.githubusercontent.com/4188758/149622996-be1d183c-4631-4c6c-b974-ffaa58399f51.png)
  - This is with Webpack (+ESBuild)
  ![Screen Shot 2022-01-15 at 11 59 22 pm](https://user-images.githubusercontent.com/4188758/149623022-c8028da2-b091-4ca8-867e-0405bab19aac.png)

- With ESBuild, it will try to do 2 things:
  - Replace the slow TS compiler in the build phase, so it can be a bit faster while running the build and minify phase. Build time has been reduced from 50 seconds to just 18 seconds.
    - This is with TS Loader:
![Screen Shot 2022-01-16 at 12 11 39 am](https://user-images.githubusercontent.com/4188758/149646729-08e8a0ad-a83c-4bb9-9800-59868c94815e.png)
    - This is with ESBuild Loader:
![Screen Shot 2022-01-16 at 2 56 51 pm](https://user-images.githubusercontent.com/4188758/149646746-853aab75-b3a0-4cd0-811c-23d71580c737.png)

  - Replace the TS test transformer from TS-jest with ESBuild-Jest. Tests will now reduce the total time to run from 15s to 5s.
    - This is with tsc 
![Screen Shot 2022-01-16 at 12 00 44 am](https://user-images.githubusercontent.com/4188758/149623053-c09acc37-48eb-4d90-aad1-c59c81352497.png)
    - This is with esbuild
![Screen Shot 2022-01-16 at 12 01 49 am](https://user-images.githubusercontent.com/4188758/149623056-296181a5-7aa7-401e-ac94-aa5738341860.png)



## Related Issue
https://github.com/viet-aus-it/vait-discord-bot/issues/79

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the **CONTRIBUTING** document.
- [ ] I have added tests to cover my changes.
- [x] All new and existing tests passed.